% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootD_equalSS.R
\name{bootD_equalSS}
\alias{bootD_equalSS}
\title{Bootstrap equal-sample-size Cohen's d}
\usage{
bootD_equalSS(x, y, nBoot = 2000)
}
\arguments{
\item{x}{The first vector}

\item{y}{The second vector}

\item{nBoot}{The number of bootstrap iterations}
}
\description{
Unequal sample sizes can skew estimates of group differences. This function
takes two numeric vectors, \code{x} and \code{y}, and repeatedly calculates 
Cohen's d for the difference in means. On each iteration the smaller of the two vectors 
is used whole, while the larger of the two vectors is randomly subsampled with the same number
of elements as the smaller vector (NAs excluded). Several 
quantiles of the resampled distributions of Cohen's d
are returned, along with the true mean difference.
}
\examples{
x <- c(NA,rnorm(3000,.5))
y <- rnorm(30)
bootD_equalSS(x,y)

}
