% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_optimizing.R
\name{brm_optimizing}
\alias{brm_optimizing}
\title{Maximum-likelihood fitting of a brm-style model}
\usage{
brm_optimizing(
  formula = NA,
  data = NA,
  ...,
  iter = 2000,
  brmModel = NA,
  strata = NA,
  crossvals = 0,
  bootstrap = NA,
  refresh = 0
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Model data}

\item{...}{Additional arguments to pass to an initial \code{brm} model}

\item{iter}{Number of iterations. If \code{strata == NA} then this is draws from a normal approximation (inverse Hessian); if \code{strata != NA} then this is the number of bootstrap samples.}

\item{brmModel}{An initial brm model to re-fit using maximum-likelihood.}

\item{strata}{The strata to use in resampling; a vector the length of the data set's rows. If all values are equal, no stratification is used in resampling}

\item{crossvals}{Number of cross-validations}

\item{bootstrap}{Bootstrapping is done by default if \code{strata} is provided. However, in certain cases (e.g., if cross-validation is desired but bootstrapping is not), bootstrapping can be turned off by setting \code{bootstrap = F}}

\item{refresh}{Should progress be printed?

Use \code{\link[rstan]{optimizing}} fit a \code{\link[brms]{brm}}-style model}
}
\description{
Maximum-likelihood fitting of a brm-style model
}
\examples{

m1 <- brm_optimizing(Sepal.Width ~ Sepal.Length * Petal.Width + (1|Species),iris)

    m_cv_1 <- brm_optimizing(Sepal.Width ~ Sepal.Length * Petal.Width + (1|Species),iris
,strata = iris$Species,iter = 500, 
bootstrap = F, crossvals = 50)
m_cv_2 <- brm_optimizing(Sepal.Width ~ Sepal.Length * Petal.Width,iris
                         ,strata = iris$Species,iter = 500, 
                         bootstrap = F, crossvals = 50)
m_cv_1$logLik_oos - m_cv_2$logLik_oos # m_cv_1 is much better! test 1, 50 crossvals: 57.7; test 2: ; test 3: 

}
