% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_cochraneEtAl_2019_PLOSOne}
\alias{dat_cochraneEtAl_2019_PLOSOne}
\title{Data from Cochrane, Simmering, & Green (2019), PLOS One}
\format{
An object of class \code{data.frame} with 122 rows and 18 columns.
}
\usage{
dat_cochraneEtAl_2019_PLOSOne
}
\description{
Data from Cochrane, Simmering, & Green (2019; DOI: 10.1371/journal.pone.0221353).
Data from child (7-8 years old) and adult participants ("subNum") completing various
cognitive tasks in a controlled lab setting.
}
\keyword{datasets}
