% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm_dr_lm.R
\name{ddm_dr_lm}
\alias{ddm_dr_lm}
\title{Fit linear predictors of DDM drift rate}
\usage{
ddm_dr_lm(formIn, data, respVar, nBoots = 2, fixBias = NA)
}
\arguments{
\item{formIn}{formula. Left hand side should be response time.}

\item{data}{data}

\item{respVar}{name of the "boundary" column (argument is character; data must be binary 0 and 1)}

\item{nBoots}{number of bootstrapped samples for CI of drift rate parameters. Must be an integer >1}

\item{fixBias}{Optional scalar value to fix the bias [beta] parameter of the Wiener function}
}
\description{
Essentially fits a generalized linear model for a Wiener process, using the density
estimation function from the \code{RWiener} package. First fits a standard 4-parameter Wiener
model to the entire RT+response-category vector. Then uses the boundary separation, bias, and
non-decision time from this overall model, and finds the best set of parameters to create a
drift rate vector as a linear function of the right-hand side of \code{formIn}. \strong{Only
numeric predictors have been tested}.
}
\details{
Note that, for parameters other than drift rate, the \code{RWiener} labeling conventions are used: 
\code{alpha} is boundary separation, \code{tau} is non-decision time, and
\code{beta} is bias.

\code{rt ~ 1} will cause an error; in principle, if a unconditional point estimate of drift rate is desired,
a dummy constant numeric variable (e.g., all 0s) could be included as a predictor.
}
\examples{
d <- data.frame(rt = .2 + exp(rnorm(50,-.5,.5))
,corr = rbinom(50,1,.5)
,totalTrialNum = 1:50
)
d$stimStr <- d$rt+rnorm(50,d$corr)
m <- ddm_dr_lm(rt ~ totalTrialNum * stimStr, data = d,respVar='corr',nBoots = 30)
m$model

}
\seealso{
\code{\link[RWiener]{wdm}} for the method upon which this is based; 
\code{\link[brms]{brm}} with family "wiener" for a much better (but slower) option.
}
