% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_Xfm.R
\name{df_xfm}
\alias{df_xfm}
\title{Transform a numeric data frame or matrix using the Yeo-Johnson algorithm}
\usage{
df_xfm(x, keepMedian = T, keepMAD = T)
}
\arguments{
\item{x}{Numeric data frame or matrix}

\item{keepMedian}{Should the result be shifted so that it has the same median as the input vector?}

\item{keepMAD}{Should the result be scaled so that it has the same dispersion (median absolute deviation) as the input vector?}
}
\description{
Returns a data frame, along with many supplementary attributes
(e.g., correlations, transformation parameters, simulate from correlation matrix).
}
\examples{
x <- data.frame(w = 'char',x=rnorm(20),y=rgamma(20,3,2),z=c(rnorm(19,3,3),NA))
x_transformed <- df_xfm(x)
head(x_transformed)
attributes(x_transformed)

}
