% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLowerRT.R
\name{findLowerRT}
\alias{findLowerRT}
\title{Find Lower RT bound, return the bound and a vector of "keep values"}
\usage{
findLowerRT(rtVector, accVector)
}
\arguments{
\item{rtVector}{vector of RTs}

\item{accVector}{vector of 0 and 1 or T and F indicating correct performance}
}
\description{
Following the recommendation of Ratcliff & Tuerlinckx (2002)
and finds the threshold at which, below that RT, responses are not different than chance
(binomial test p-value > .05). RT thresholds are tested in steps of .001 if RT is 
in seconds, steps of 1 if RT is in milliseconds. Begins by rejecting the lowest
10 RTs to start the binomial tests. If the number rejected is below 10 or so, 
consider not rejecting any! Returns lowerRT,
the rejection level, as well as keepVect, a vector tagging which RTs to keep.
}
