% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitQuickfun.R
\name{fitQuickfun}
\alias{fitQuickfun}
\title{Fit a Quick [see Ahissar & Hochstein] function, given data [currently nonsense fits]}
\usage{
fitQuickfun(
  x,
  y,
  t = NA,
  model = "static",
  lapseRate = 0.02,
  numGuesses = 25,
  maxThresh = 5
)
}
\arguments{
\item{x}{stimulus strengths to fit [positive, e.g., SOA]}

\item{y}{responses to fit [bernoulli]}

\item{t}{vector of trial numbers, if desired}

\item{model}{"static" or "exp," depending on what the parameters correspond to}

\item{lapseRate}{At what lapse rate were the parameters fit?}

\item{maxThresh}{How high of a threshold is acceptable (in multiples of the largest stimulus strength value)?}
}
\description{
Given data, what are the predicted Quick parameters?
Fits [by minimizing negative log likelihood] either a time-evolving or a static Quick function 
(Ahissar & Hochstein). If time-evolving,
then threshold is each fit by a 3-parameter exponential decay function.
}
\details{
Returns a data frame with fit parameters and the [negative] log likelihood.
}
\examples{
example
}
\keyword{keywords}
