% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTime.R
\name{getTime}
\alias{getTime}
\title{Get the current date and time in an easy-to-use format}
\usage{
getTime(style = 1)
}
\arguments{
\item{style}{style 1 uses letters to convey the month, style 2 uses numbers to convey the month.}
}
\description{
Returns a string with the date and time.
}
\examples{

getTime()
getTime(style = 2)

}
\keyword{keywords}
