% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertYJ.R
\name{invertYJ}
\alias{invertYJ}
\title{Invert a Yeo-Johnson transformation}
\usage{
invertYJ(xt, lambda, interval = c(-1e+10, 1e+10), keepMedian = T, keepMAD = T)
}
\arguments{
\item{xt}{Transformed values of data}

\item{lambda}{Lambda value that had transformed data into xt}

\item{interval}{Interval in which to look for data. If search is difficult, restricting the search to plausible values should help.}

\item{keepMedian}{Logical. Should the original data's median be kept?}

\item{keepMAD}{Logical. Should the original data's MAD be kept?}
}
\description{
Invert a Yeo-Johnson transformation
}
\examples{
x <- 1:10
x1 <- VGAM::yeo.johnson(x,2)
x_recovered <- invertYJ(x1,2)

d <- rnorm(5,3) + exp(5)
d1 <- AaronFuns::YeoJohn(d)
invertYJ(attr(d1,'transformedDat'),attr(d1,'lambda'),interval = c(-1e3,1E3)) # fails with too large of an interval.
}
