% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isOutlier.R
\name{isOutlier}
\alias{isOutlier}
\title{Find multivariate outliers (Leys et al., 2018, JESP)}
\usage{
isOutlier(dat, dat_proportion = "check", reject_alpha = 0.01)
}
\arguments{
\item{dat}{data}

\item{dat_proportion}{Proportion of data to retain in estimating the robust multivariate location and scatter. Defaults to testing at various proportions.}

\item{reject_alpha}{P-value threshold for excluding cases (using a chi-square test on Mahalanobis distances; Leys et al, 2018)}
}
\description{
Tests for multivariate outliers using a robust method (cited above). In summary, a minimum proportion of
cases are used to estimate the location and scatter of the multivariate space, then that location and scatter are
used to test for [multivariate] outlying values.
}
\details{
Takes a numeric data frame. Returns a logical vector indicating whether the cases should be excluded.

Is \emph{NOT} robust to missing data.
}
\seealso{
\code{\link[MASS]{cov.rob}} for estimating the robust location and scatter.
}
