% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer2table.R
\name{lmer2table}
\alias{lmer2table}
\title{Put lmer results into a data frame}
\usage{
lmer2table(modIn, modID1 = NA, modID2 = NA, modID3 = NA, round2 = 4)
}
\arguments{
\item{modIn}{lmer object}

\item{modID1}{optional identification variable (e.g., dataframe name)}

\item{modID2}{optional identification variable (e.g., task name)}

\item{modID3}{optional identification variable (e.g., a grouping variable)}
}
\description{
Input an lmer object, and up to three optional descriptions. In return, you get a table
with the model's estimates of F, p, df, b, and CIs.
}
\details{
Unfortunately, this also creates a modelOutputByproducts.txt file... Which is a compromise,
to prevent lmSupport from blowing up the console with a ton of output.
}
\keyword{keywords}
