% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_cohenD.R
\name{lmer_cohenD}
\alias{lmer_cohenD}
\title{Find Cohen's D for lmer model}
\usage{
lmer_cohenD(lmerMod)
}
\arguments{
\item{lmerMod}{Model fit by \code{lme4::lmer()}}
}
\description{
Calculate a heuristic Cohen's D for each fixed effect in an lmer model.
Uses a fairly naive calculation: \code{abs(b)/(SE*sqrt(df))}, where df is approximated
using a Kenward-Rogers approximation (package \code{pbkrtest} via \code{lmSupport::modelSummary}). 
Returns the \code{lmSupport::modelSummary} output with the addition of \code{$cohen_d}.
}
\details{
As with (& because of the underlying K-R approximation in) \code{lmSupport::modelSummary},
this can take a very long time for large or complex models.
}
