% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oos_partial.R
\name{oos_partial}
\alias{oos_partial}
\title{Estimate matrix of bivariate out-of-sample variance explained}
\usage{
oos_partial(xvars, partialvars = c(), data, nBoots = 200)
}
\arguments{
\item{xvars}{Vector of variable names, with which to estimate bivariate relations.}

\item{partialvars}{If a vector of variable names are provided, then the lower triangle of the output is the out-of-sample variance explained \emph{when controlling for the variables in \code{partialvars}}.}

\item{data}{Data to use}

\item{nBoots}{Number of resamples}
}
\value{

}
\description{
Uses \code{tef_rlm_boot()} to get out-of-sample variance explained as well as bootstrapped p values.
The latter is two times the proportion of the bootstrapped distribution is on the opposite side of
zero than the median coefficient.
}
