% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsplot.R
\name{pairsplot}
\alias{pairsplot}
\title{Create a \code{psych::pairs.panels()} plot with standard formatting}
\usage{
pairsplot(d, method = "Spearman", main = "", stars = T, color = "darkgrey")
}
\arguments{
\item{d}{Data frame to be plotted}

\item{method}{Correlation method to be used: product-moment ('pearson') or rank ('spearman') ?}

\item{main}{Plot title, as in plot()}

\item{stars}{Do you want magic stars? (same as \code{pairs.panels()})}

\item{color}{What color do you want in the histogram bars?}
}
\description{
The default title includes approximate Bayes Factors-based estimates of regions providing "At least moderate
evidence for the lack of a correlation" or "at least moderate evidence for a correlation." The BF approximation
uses the \code{BayesFactor::ttest.tstat()} function, and finds the r value associated with a T value associated with a
raw BF <.33333 or >3. \strong{Beware}: The \emph{n} used in these calculations is the number of rows of the data frame.
If you have missing data, these estimates could be wildly off.
}
\examples{
d <- iris[1:100,c('Sepal.Length','Sepal.Width','Petal.Length','Petal.Width')]
d$isoDist <- ACmisc::d1iso(d)
pairsplot(d)

}
