% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustCor.R
\name{robustCor}
\alias{robustCor}
\title{Find a robust bootstrapped correlation}
\usage{
robustCor(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
First optimally transforms x and y using \code{\link{YeoJohn}} (i.e., Yeo-Johnsons transformation
with a lambda optimized to minimize skew).
Then resamples the data 2000 times and, for each resample, calculates the
Pearson correlation coefficient along with the corresponding T value and
log-3 Bayes Factor (calculated with a default \code{BayesFactor::correlationBF(x,y,'medium')}).
Returns a formatted string \code{$results},
the full resampled set of values in \code{$resampled_tests}, as well as
individual values for the median \code{$r} and its corresponding \code{$t} and log-2 \code{$bf}.
The \code{$data} is returned in a data frame with variables \code{xt} [transformed x] and
\code{yt} [transformed y], each of which has a set of attributes (e.g., the "rawDat" and the Yeo-Johnson "lambda").
}
\details{
The 'medium' prior for \code{BayesFactor::correlationBF} is a transformed beta(3,3) (
see the BayesFactor documentation as well as Ly, Verhagen, and Wagenmakers (2015) )

Note that this function will not work with fewer than 9 cases. Because you probably shouldn't be
calculating correlations with only 8 cases.
}
