% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale2.R
\name{scale2}
\alias{scale2}
\title{Scale a variable to a new mean and standard deviation}
\usage{
scale2(x, newmean = 0, newsd = 1, yj = F)
}
\arguments{
\item{x}{Data to scale}

\item{newmean}{Desired mean}

\item{newsd}{Desired standard deviation}

\item{yj}{[logical] Whether Yeo-Johnson transformation should be applied to minimize univariate skew}
}
\description{
Scale a variable to a new mean and standard deviation
}
\examples{

y <- c(rnorm(200,3,3),rnorm(100,10,3) )
hist(y, breaks = 40)
hist(scale2(y,4,2), breaks = 40)
hist(scale2(y,-1,.5), breaks = 40)
hist(scale2(y,10,1,yj=T),breaks = 40)
hist(scale2(exp(y),10,1,yj=F),breaks = 40)
hist(scale2(exp(y),10,1,yj=T),breaks = 40)

}
