% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterMat.R
\name{scatterMat}
\alias{scatterMat}
\title{Create a matrix of scatter plots, with each row and each variable containing one variable}
\usage{
scatterMat(df, colNames, rowNames, corMethod = "spearman", lineColor = "black")
}
\arguments{
\item{df}{Data frame}

\item{colNames}{Character vector of variable names to create the columns of the correlation matrix}

\item{rowNames}{Character vector of variable names to create the rows of the correlation matrix}

\item{corMethod}{Should the reported correlation be "pearson" or "spearman"?}
}
\description{
Create a matrix of scatter plots, with each row and each variable containing one variable
}
\examples{

scatterMat(iris, colNames = c('Sepal.Width','Petal.Width'),c('Sepal.Length','Petal.Length'))

}
