% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceIfChanged.R
\name{sourceIfChanged}
\alias{sourceIfChanged}
\title{Source a file, but only if it changed from the last time you ran \emph{sourceIfChanged()} on the file}
\usage{
sourceIfChanged(filepath)
}
\arguments{
\item{filepath}{.R file to source

The intended use is largely for when the user wants reproducible results (sometimes requiring
long analyses) while writing .Rmarkdown documents. While working with the main text, formatting,
and such, it's nice to just load pre-run data rather than waiting to source every time.

There are important implications, such as the fact that only the \emph{sourced file} is checked for
changes, and therefore the old data will be loaded even if other things (e.g., global variables)
have been changed.}
}
\description{
Rather than sourcing your analysis script many times, which can be cumbersome, this script skips the
sourcing \emph{if} there is a \code{.log} file that matches the sourced file. If that matching \code{.log}
file exists, the corresponding \code{.RData} file is loaded. Otherwise, the file is sourced and both
\code{.log} and \code{.RData} files are created.
}
