% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialWDM.R
\name{trialWDM}
\alias{trialWDM}
\title{Find by-trial Wiener Diffusion Model parameters}
\usage{
trialWDM(
  dat,
  fit_DR = T,
  fit_BS = T,
  fit_NDT = T,
  fit_Bias = F,
  winsorize = NA
)
}
\arguments{
\item{dat}{Response time vector. Positive values indicates upper-boundary while negative values indicate lower-boundary.}

\item{fit_DR}{Logical or numeric scalar. If \code{TRUE} then drift rate is estimated for each trial, if \code{FALSE} only the overall parameter is estimated, and if numeric then the parameter is fixed to that value.}

\item{fit_BS}{Logical or numeric scalar. If \code{TRUE} then boundary separation is estimated for each trial, if \code{FALSE} only the overall parameter is estimated, and if numeric then the parameter is fixed to that value.}

\item{fit_NDT}{Logical or numeric scalar. If \code{TRUE} then non-decision time is estimated for each trial, if \code{FALSE} only the overall parameter is estimated, and if numeric then the parameter is fixed to that value.}

\item{fit_Bias}{Logical or numeric scalar. If \code{TRUE} then bias is estimated for each trial, if \code{FALSE} only the overall parameter is estimated, and if numeric then the parameter is fixed to that value.}

\item{winsorize}{Optional numeric scalar. If supplied, then all univariate values greater than number of MAD from the median are changed to be instead that number of MAD from the median (using \code{mad()} defaults).}
}
\description{
Estimates a set of Wiener Diffusion Model parameters for a vector of formatted
data. Estimation uses \code{\link{RWiener::wdm}} and data must be formatted as a 
1-dimensional vector according the to \code{RWiener} package (i.e., RT * 2 
* (boundary=='upper' - 0.5) ).
}
\examples{
\dontrun{
dat <- (rnorm(200,.2,.05) + rexp(200,1 / .2))*sample(c(-1,1,1),200,replace=T)
    m <- trialWDM(dat)
    
    m2 <- trialWDM(dat, fit_NDT = F , winsorize = 2)
    
    }
}
