% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcZ.R
\name{wilcZ}
\alias{wilcZ}
\title{Wilcoxon tests, returning a z value that respects the sign of the differences between two vectors}
\usage{
wilcZ(x, y, pairTest = F)
}
\arguments{
\item{x}{First vector to compare}

\item{y}{Second vector to compare}

\item{pairTest}{Are x and y paired?}
}
\description{
Essentially, a wrapper for \code{ wilcox.test} that also returns the p value, 
Z value, W value, and the difference in medians.
}
\examples{
# Note that this doesn't make *a lot* of sense, since they're two totally different measurements
wilcZ(dat_cochraneEtAl_2019_PLOSOne$ufov, dat_cochraneEtAl_2019_PLOSOne$mot,pairTest = T)

}
