% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YeoJohn.R
\name{YeoJohn}
\alias{YeoJohn}
\title{Apply optimal Yeo-Johnson transformation to a numeric vector}
\usage{
YeoJohn(x, keepMedian = T, keepMAD = T)
}
\arguments{
\item{x}{numeric vector to transform}

\item{keepMedian}{Should the result be shifted so that it has the same median as the input vector?}

\item{keepMAD}{Should the result be scaled so that it has the same dispersion (median absolute deviation) as the input vector?}
}
\description{
Finds the Yeo-Johnson transformation that minimizes the skewness of the vector.
Returns the transformed vector with an attribute "lambda" (transformation parameter).
}
