% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroOneNorm.R
\name{zeroOneNorm}
\alias{zeroOneNorm}
\title{Normalize a vector to be between .001 and .999}
\usage{
zeroOneNorm(dat)
}
\arguments{
\item{dat}{numeric vector to normalize}
}
\description{
To force a variable to have an easily interpretable min and max.
}
\examples{
examples
zeroOneNorm(rnorm(30))
zeroOneNorm(dat_cochraneEtAl_2019_PLOSOne$mot)
}
\keyword{keywords}
