% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1iso.R
\name{d1iso}
\alias{d1iso}
\title{Extract first dimension using nonmetric scaling}
\usage{
d1iso(data, dist_method = c("manhattan", "euclidean"))
}
\arguments{
\item{data}{Must be numeric}

\item{dist_method}{'manhattan' (default; L1 norm) or 'euclidean' (AKA OLS or L2 norm)}
}
\description{
Implements \code{\link[MASS]{isoMDS}(dist(data,dist_method),k=1)}, with
attributes of correlations with (scaled) rowMeans and with PCA component-1.
}
