% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2bf.R
\name{r2bf}
\alias{r2bf}
\title{Convert correlation \emph{r} to a Bayes Factor}
\usage{
r2bf(r, n)
}
\arguments{
\item{r}{}

\item{n}{}
}
\description{
Uses default values from the \code{BayesFactor::ttest.tstat} function to convert an 
effect size \emph{r} to a Bayes factor, given the sample size \emph{n}. Also includes as attributes the thresholds, 
at that sample size, for conventionally "accepting the null hypothesis" (i.e., BF < .333)
or "rejecting the null hypothesis" (i.e., BF > 3). Bayes Factors are returned "raw," that is,
as the ratio of the evidence for the alternative hypothesis to the evidence for the null hypothesis.
}
\examples{
exampleR <- cor(dat_cochraneEtAl_2019_PLOSOne$enum , dat_cochraneEtAl_2019_PLOSOne$mot)
r2bf(exampleR , nrow(dat_cochraneEtAl_2019_PLOSOne))

}
