% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2bf.R
\name{r2bf}
\alias{r2bf}
\title{Convert correlation \emph{r} to a Bayes Factor}
\usage{
r2bf(r, n, rscale = "medium", nsims = 20, returnThresholds = FALSE)
}
\arguments{
\item{r}{}

\item{n}{}

\item{rscale}{Passes to the \code{rscale} argument of \code{\link[BayesFactor]{ttest.tstat}}. Choose from "medium.narrow", "medium", "wide", and "ultrawide". As in \code{ttest.tstat}, defaults to "medium".}

\item{nsims}{Not currently used.}

\item{returnThresholds}{Logical. Should the boundaries indicating "decision thresholds" be estimated?}
}
\description{
Uses default values from the \code{\link[BayesFactor]{ttest.tstat}} function to convert an 
effect size \emph{r} to a Bayes factor by first finding the T value, given the sample size \emph{n}. 
Can also include as attributes the thresholds, 
at that sample size, for conventionally "accepting the null hypothesis" (i.e., BF < .333)
or "rejecting the null hypothesis" (i.e., BF > 3). Bayes Factors are returned "raw," that is,
as the ratio of the evidence for the alternative hypothesis to the evidence for the null hypothesis.
All values returned should be treated as \strong{approximate}, and the hypothesis 
acceptance thresholds should be treated as \strong{heuristics}. For Bayes Factors for
a particular correlation, using \code{\link[BayesFactor]{correlationBF}} is \strong{highly recommended}.
}
\examples{
exampleR <- cor(dat_cochraneEtAl_2019_PLOSOne$enum , dat_cochraneEtAl_2019_PLOSOne$mot)
r2bf(exampleR , nrow(dat_cochraneEtAl_2019_PLOSOne))



}
