% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Find Root Mean Squared Error}
\usage{
rmse(vect1, vect2)
}
\arguments{
\item{vect1}{A numeric vector}

\item{vect2}{Another numeric vector, of the same length as \code{vect1}}
}
\description{
Simply finds the RMSE between vect1 and vect2 (of the same length). 
Often, when assessing fit values, this is interpreted as the standard 
deviation of fits (e.g., vect2) around true values (e.g., vect1).
}
