% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyReg.R
\name{shinyReg}
\alias{shinyReg}
\title{Visualize two-predictor regression using shiny}
\usage{
shinyReg(dat, overallTitle = "Data Exploration")
}
\arguments{
\item{dat}{data frame that you would like to explore}

\item{overallTitle}{What to you want your window's big title to be?}
}
\description{
Opens shiny window to allow basic interactive exploration of a dataset.
}
\examples{
shinyReg(dat_cochraneEtAl_2019_PLOSOne)

}
