% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_from_dat.R
\name{sim_from_dat}
\alias{sim_from_dat}
\title{Simulate new data from a dataset's quantiles}
\usage{
sim_from_dat(nSim, datIn, nTries = 500)
}
\arguments{
\item{nSim}{Number of simulated observations}

\item{datIn}{Data to simulate from}

\item{nTries}{Number of simulated datasets to generate. Only one, the one with the rank correlation closest to the original dataset's, will be returned.

Given an existing dataset's rank correlation structure and the quantiles of 
the variables in that dataset, a new dataset is simulated from the quantiles
and attempting to match the rank correlation of the original.}
}
\description{
Simulate new data from a dataset's quantiles
}
\examples{

new_iris <- sim_from_dat(200,iris) # note that non-numeric variables will be dropped

}
\seealso{
\code{\link{sim_dat}} accomplishes something similar, although that function resamples directly from the
original data while this one resamples from the quantiles of the original data.
}
