% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcZ.R
\name{wilcZ}
\alias{wilcZ}
\title{Wilcoxon tests, returning a z value that respects the sign of the differences between two vectors}
\usage{
wilcZ(x, y, pairTest = F)
}
\arguments{
\item{x}{First vector to compare}

\item{y}{Second vector to compare}

\item{pairTest}{Are x and y paired?}
}
\description{
Also returns the p value, W value, and the difference in medians.
}
